+++ crates/macros/tests/mods_raw_branch/_.rs
@@ -1,0 +1,3 @@
+//!
+///
+pub const VALUE: u8 = 4;