+++ docs/changelog.md
@@ -76,8 +76,11 @@
   - add indexed slice iteration.
   - add forward range continuation with `i in ..end` and `i in ..=end`.
 - update `doclink!` and `_reexport!`: support linking to the item directly.
-- update `structural_mods!`: remove the `_workspace_internals` branch.
 - update `test_size_of!`: support negative niche `Option` testing.
+
+#### code::util::synth
+- new macro: `mods_in!`.
+- update `structural_mods!`: remove the `_workspace_internals` branch.
 
 ### data::access
 - new type `StridedBlocks`.